Simplify shutdown by removing sendCh ref tracking
Previously, the sendCh was closed, we had to carefully track all
references to it, as any send to a closed channel would panic.

However, this was finally removed in #388:
https://github.com/uber/tchannel-go/pull/388/commits/93ef5c112c8b321367ae52d2bd79396e2e874f31

As mentioned in the change, we can simplify the shutdown path
as we no longer need to track references to sendCh. Caller can
attempt writes to sendCh, but those frames will not be sent
as they race with the close.